# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aura social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: getFeed
      tags: [videos]
      summary: Get paginated video feed
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
          description: Last seen video ID for cursor pagination
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [for_you, following]
          description: Feed mode — "following" returns only videos from followed users (requires auth)
      responses:
        "200":
          description: Video feed page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPage"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a video record after upload
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoBody"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete own video
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos/{id}/view:
    post:
      operationId: recordView
      tags: [videos]
      summary: Record a view on a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: View recorded
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Video:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        videoPath:
          type: string
        thumbnailPath:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        status:
          type: string
          enum: [processing, ready, failed]
        viewsCount:
          type: integer
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
          description: Whether the current user liked this video (false if unauthenticated)
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/VideoAuthor"
      required:
        - id
        - userId
        - videoPath
        - status
        - viewsCount
        - likesCount
        - commentsCount
        - isLiked
        - createdAt
    VideoAuthor:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - username
        - displayName
        - avatarUrl
    FeedPage:
      type: object
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        nextCursor:
          type: ["integer", "null"]
      required:
        - videos
        - nextCursor
    CreateVideoBody:
      type: object
      properties:
        videoPath:
          type: string
        thumbnailPath:
          type: string
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        duration:
          type: integer
      required:
        - videoPath
...